requirements.txt
# Core Backend
fastapi>=0.108.0
uvicorn>=0.27.0
sqlalchemy>=2.0.0
pydantic>=2.5.0
alembic>=1.13.0
psycopg2-binary>=2.9.9
pgvector>=0.2.4
python-dotenv>=1.0.0
httpx>=0.25.0

# Document Processing
pypdf>=4.0.0
docx2txt>=0.8
openpyxl>=3.1.2
markdown>=3.5.0
python-pptx>=0.6.21
beautifulsoup4>=4.12.0
html2text>=2020.1.16

# Utilities
tqdm>=4.66.0
pytest>=7.4.3
tenacity>=8.2.3
loguru>=0.7.2